added SLOPE to penalty submodule
