+++ skglm/penalties/__init__.py
@@ -1,6 +1,6 @@
 from .base import BasePenalty
 from .separable import (
-    L1_plus_L2, L0_5, L1, L2_3, MCPenalty, SCAD, WeightedL1, IndicatorBox
+    L1_plus_L2, L0_5, L1, L2_3, MCPenalty, SCAD, WeightedL1, IndicatorBox, SLOPE
 )
 from .block_separable import (
     L2_05, L2_1, BlockMCPenalty, BlockSCAD, WeightedGroupL2
@@ -9,6 +9,6 @@ from .block_separable import (
 
 __all__ = [
     BasePenalty,
-    L1_plus_L2, L0_5, L1, L2_3, MCPenalty, SCAD, WeightedL1, IndicatorBox,
+    L1_plus_L2, L0_5, L1, L2_3, MCPenalty, SCAD, WeightedL1, IndicatorBox, SLOPE,
     L2_05, L2_1, BlockMCPenalty, BlockSCAD, WeightedGroupL2
 ]
